+++ MBW.Server/Controllers/ReplyController.cs
@@ -2,6 +2,7 @@
 using System.Data.Common;
 using System.Security.Claims;
 using MBW.Server.DTO;
+using MBW.Server.Enum;
 using MBW.Server.Models;
 using MBW.Server.Utils;
 using Microsoft.AspNetCore.Authorization;
@@ -67,18 +68,20 @@ public class ReplyController : ControllerBase
     {
         try
         {
-            User? user = await _dbContext.Users.FirstOrDefault(User.FindFirst(ClaimTypes.Name)?.Value)
-            // NEEDS USER VALIDATION
+            User? user = await _dbContext.Users.FirstOrDefaultAsync(u => u.Name == User.FindFirst(ClaimTypes.Name).Value);
             Reply? res = _dbContext.Replies.FirstOrDefault(r => r.Id == reply.Id);
 
             if (res == null)
                 return NotFound();
+            
+            if (user == null || (user.Id != res.UserId && user.Role != Roles.ADMIN))
+                return Unauthorized("This is not your post");
 
             res.Content = reply.Content;
             _dbContext.Replies.Update(res);
-            await _dbContext.SaveChangesAsync().ConfigureAwait(false);
+            await _dbContext.SaveChangesAsync();
             
-            return Ok(res); // 200 Ok
+            return Ok(res);
         }
         catch (DbException)
         {
@@ -93,16 +96,19 @@ public class ReplyController : ControllerBase
     {
         try
         {
-            // NEEDS USER VALIDATION
+            User? user = await _dbContext.Users.FirstOrDefaultAsync(u => u.Name == User.FindFirst(ClaimTypes.Name).Value);
             Reply? res = _dbContext.Replies.FirstOrDefault(r => r.Id == replyId);
-            
+
             if (res == null)
-                return NoContent(); // 204 No Content
+                return NotFound();
+            
+            if (user == null || (user.Id != res.UserId && user.Role != Roles.ADMIN))
+                return Unauthorized("This is not your post");
             
             _dbContext.Replies.Remove(res);
-            await _dbContext.SaveChangesAsync().ConfigureAwait(false);
+            await _dbContext.SaveChangesAsync();
             
-            return Ok(res); // 200 Ok
+            return Ok(res);
         }
         catch (DbException)
         {
